% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{entry.pvalue}
\alias{entry.pvalue}
\title{Compute entrywise p-values for DrFARM}
\usage{
entry.pvalue(X, Y, Theta, B, E.Z, precM, standardize = TRUE)
}
\arguments{
\item{X}{A size \eqn{n \times p} matrix of predictors (e.g., genetic variants). Missing values are not allowed}

\item{Y}{A size \eqn{n \times q} matrix of outcomes (e.g., continuous omics traits). Missing values are not allowed}

\item{Theta}{A size \eqn{q \times p} initial coefficient matrix}

\item{B}{A size \eqn{q \times k} factor loading matrix}

\item{E.Z}{A size \eqn{n \times k} latent factor score matrix}

\item{precM}{A size \eqn{p \times p} precision matrix}

\item{standardize}{A logical indicating whether to standardize \code{X} and \code{Y} by column. Default is \code{TRUE}}
}
\value{
A size \eqn{q \times p} matrix of p-values
}
\description{
Computes entrywise p-values for DrFARM derived from outer-debiasing
}
