% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{precM.glasso}
\alias{precM.glasso}
\title{Estimate precision matrix via \code{glasso}}
\usage{
precM.glasso(
  X,
  standardize = TRUE,
  gamma = 0.5,
  rholist = NULL,
  thr = 1e-04,
  maxit = 10000
)
}
\arguments{
\item{X}{A size \eqn{n \times p} matrix of predictors (e.g., genetic variants). Missing values are not allowed}

\item{standardize}{A logical indicating whether to standardize \code{X} by column. Default is \code{TRUE}}

\item{gamma}{A numeric value in \([0, 1]\) for the EBIC hyperparameter. When \code{gamma = 0}, EBIC reduces to the ordinary BIC. Default is \code{0.5}}

\item{rholist}{A list of nonnegative regularization parameters for the lasso
(penalizing off-diagonal entries in the \code{glasso} framework).}

\item{thr, maxit}{Numerics specifying convergence threshold and maximum iterations for \code{glasso},
respectively. Defaults are \code{1e-4} and \code{1e4}.}
}
\value{
A size \eqn{p \times p} precision matrix estimate
}
\description{
Searches over a grid of tuning parameters in \code{rholist}, computes \code{EBIC} for each, and
returns the \eqn{p x p} precision matrix that minimizes the EBIC
}
