% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{precM.NL.whole}
\alias{precM.NL.whole}
\title{Estimate whole precision matrix via Nodewise Lasso}
\usage{
precM.NL.whole(
  X,
  standardize = TRUE,
  method = "EBIC",
  gamma = 0.5,
  n.lambda = 100
)
}
\arguments{
\item{X}{A size \eqn{n \times p} matrix of predictors (e.g., genetic variants). Missing values are not allowed}

\item{standardize}{A logical indicating whether to standardize \code{X} by column. Default is \code{TRUE}}

\item{method}{A character, either \code{"EBIC"} or \code{"CV"}, determining how \code{lambda} is chosen}

\item{gamma}{A numeric value in \([0, 1]\) for the EBIC hyperparameter. When \code{gamma = 0}, EBIC reduces to the ordinary BIC. Default is \code{0.5}}

\item{n.lambda}{An integer giving the number of \code{lambda} points for \code{glmnet}. Default is \code{100}}
}
\value{
A \eqn{p \times p} precision matrix estimate
}
\description{
Estimates the entire precision matrix via Nodewise Lasso by minimizing \code{EBIC}
or via cross-validation (\code{CV}).
}
