% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{precM.QO}
\alias{precM.QO}
\title{Estimate precision matrix via Quadratic Optimization}
\usage{
precM.QO(X, standardize = TRUE)
}
\arguments{
\item{X}{A size \eqn{n \times p} matrix of predictors (e.g., genetic variants). Missing values are not allowed}

\item{standardize}{A logical indicating whether to standardize \code{X} by column. Default is \code{TRUE}}
}
\value{
A size \eqn{p \times p} precision matrix
}
\description{
Estimate precision matrix via Quadratic Optimization
}
