% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{remMap.grid}
\alias{remMap.grid}
\title{Generate tuning parameter grid for \code{remMap}}
\usage{
remMap.grid(X, Y, standardize = TRUE, n.lambda = 10, lambda.min.ratio = 0.01)
}
\arguments{
\item{X}{A size \eqn{n \times p} matrix of predictors (e.g., genetic variants). Missing values are not allowed}

\item{Y}{A size \eqn{n \times q} matrix of outcomes (e.g., continuous omics traits). Missing values are not allowed}

\item{standardize}{A logical indicating whether to standardize \code{X} and \code{Y} by column. Default is \code{TRUE}}

\item{n.lambda}{The number of lambda points in a single dimension. Default is \code{10}}

\item{lambda.min.ratio}{The smallest value for lambda as a fraction of \code{lambda.max}. Default is \code{0.01}}
}
\value{
A data frame with two columns: \code{lambda1} and \code{lambda2}
}
\description{
Generates a 2D grid of tuning parameters (\code{lambda1}, \code{lambda2}) for \code{remMap}
}
