% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{DrFARM.grid}
\alias{DrFARM.grid}
\title{Generate tuning parameter grid for \code{DrFARM}}
\usage{
DrFARM.grid(
  X,
  Y,
  Theta0,
  precM,
  k,
  lambda1.opt,
  lambda2.opt,
  K = NULL,
  standardize = TRUE
)
}
\arguments{
\item{X}{A size \eqn{n \times p} matrix of predictors (e.g., genetic variants). Missing values are not allowed}

\item{Y}{A size \eqn{n \times q} matrix of outcomes (e.g., continuous omics traits). Missing values are not allowed}

\item{Theta0}{A size \eqn{q \times p} initial coefficient matrix}

\item{precM}{A size \eqn{p \times p} precision matrix}

\item{k}{A positive integer specifying the number of latent factors used in DrFARM}

\item{lambda1.opt}{The chosen lasso tuning parameter from \code{remMap.whole}}

\item{lambda2.opt}{The chosen group-lasso tuning parameter from \code{remMap.whole}}

\item{K}{An optional size \eqn{n \times n} kinship matrix. Default is \code{NULL}}

\item{standardize}{A logical indicating whether to standardize \code{X} and \code{Y} by column. Default is \code{TRUE}}
}
\value{
A data frame with two columns: \code{lambda1} and \code{lambda2}
}
\description{
Generates a 2D grid of tuning parameters (\code{lambda1}, \code{lambda2}) for \code{DrFARM}
}
