% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{DrFARM.one}
\alias{DrFARM.one}
\title{Fit one \code{DrFARM} model}
\usage{
DrFARM.one(
  X,
  Y,
  Theta0,
  precM,
  k,
  lambda1,
  lambda2,
  K = NULL,
  C = NULL,
  standardize = TRUE,
  thres = 1e-04,
  rotate = "none",
  scores = "regression",
  fm = "ml"
)
}
\arguments{
\item{X}{A size \eqn{n \times p} matrix of predictors (e.g., genetic variants). Missing values are not allowed}

\item{Y}{A size \eqn{n \times q} matrix of outcomes (e.g., continuous omics traits). Missing values are not allowed}

\item{Theta0}{A size \eqn{q \times p} initial coefficient matrix}

\item{precM}{A size \eqn{p \times p} precision matrix}

\item{k}{A positive integer specifying the number of latent factors used in DrFARM}

\item{lambda1}{A user supplied value for the lasso tuning parameter}

\item{lambda2}{A user supplied value for the group lasso tuning parameter}

\item{K}{An optional size \eqn{n \times n} kinship matrix. Default is \code{NULL}}

\item{C}{A size \eqn{q \times p} integer matrix specifying model inclusion and penalty rules:
\itemize{
  \item 0 = excluded from model
  \item 1 = penalized
  \item 2 = unpenalized
}
If \code{null}, all entries are penalized (1) by default}

\item{standardize}{A logical indicating whether to standardize \code{X} and \code{Y} by column. Default is \code{TRUE}}

\item{thres}{A numeric value for the convergence threshold. Default is \eqn{1 \times 10^{-4}}}

\item{rotate}{Character string specifying factor analysis options, consistent with the \pkg{psych} package (e.g., \code{"none"})}

\item{scores}{Character string specifying factor analysis options, consistent with the \pkg{psych} package (e.g., \code{"regression"})}

\item{fm}{Character string specifying factor analysis options, consistent with the \pkg{psych} package (e.g., \code{"ml"})}
}
\value{
A list containing:
\item{Theta}{A size \eqn{q \times p} coefficient matrix}
\item{B}{A size \eqn{q \times k} factor loading matrix}
\item{E.Z}{A size \eqn{n \times k} latent factor score matrix}
\item{diag.Psi}{A length-\eqn{q} vector of uniquenesses (diagonal of \eqn{\Psi}).}
}
\description{
Fit \code{DrFARM} for a single pair (\code{lambda1}, \code{lambda2}), serving as a helper function for
\code{DrFARM.all} to facilitate for parallelization.
}
