% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{remMap.EBIC}
\alias{remMap.EBIC}
\title{Compute EBIC for \code{remMap}}
\usage{
remMap.EBIC(X, Y, Theta0, standardize = TRUE, gamma = 1)
}
\arguments{
\item{X}{A size \eqn{n \times p} matrix of predictors (e.g., genetic variants). Missing values are not allowed}

\item{Y}{A size \eqn{n \times q} matrix of outcomes (e.g., continuous omics traits). Missing values are not allowed}

\item{Theta0}{A \eqn{q \times q} coefficient matrix}

\item{standardize}{A logical indicating whether to standardize \code{X} and \code{Y} by column. Default is \code{TRUE}}

\item{gamma}{A numeric value in \([0, 1]\) for the EBIC hyperparameter. When \code{gamma = 0}, EBIC reduces to the ordinary BIC. Default is \code{1}}
}
\value{
A numeric EBIC value
}
\description{
Computes the Extended Bayesian Information Criterion (EBIC) for a given \eqn{\Theta0}.
}
