% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{remMap.one}
\alias{remMap.one}
\title{Fit one \code{remMap} model}
\usage{
remMap.one(X, Y, standardize = TRUE, lambda1, lambda2, C = NULL)
}
\arguments{
\item{X}{A size \eqn{n \times p} matrix of predictors (e.g., genetic variants). Missing values are not allowed}

\item{Y}{A size \eqn{n \times q} matrix of outcomes (e.g., continuous omics traits). Missing values are not allowed}

\item{standardize}{A logical indicating whether to standardize \code{X} and \code{Y} by column. Default is \code{TRUE}}

\item{lambda1}{A user supplied value for the lasso tuning parameter}

\item{lambda2}{A user supplied value for the group lasso tuning parameter}

\item{C}{A size \eqn{q \times p} integer matrix specifying model inclusion and penalty rules:
\itemize{
  \item 0 = excluded from model
  \item 1 = penalized
  \item 2 = unpenalized
}
If \code{null}, all entries are penalized (1) by default}
}
\value{
A \eqn{q \times p} coefficient matrix, \code{Theta0}
}
\description{
Fit \code{remMap} for a single pair (\code{lambda1}, \code{lambda2}), serving as a helper function for
\code{remMap.all} to facilitate for parallelization.
}
